chore: separated different aspects of TrendingCardsViewController
