+++ Headlines/src/Controllers/TrendingCardsViewController.swift
@@ -9,12 +9,8 @@
 import UIKit
 import SDWebImage
 import Crashlytics
-import StoreKit
 
-class TrendingCardsViewController: UIViewController,
-                                    UICollectionViewDelegate,
-                                    UICollectionViewDataSource,
-                                    TabbedViewController {
+class TrendingCardsViewController: UIViewController {
 
     var topics = [Topic]()
     var footerView: UICollectionReusableView?
@@ -140,28 +136,6 @@ class TrendingCardsViewController: UIViewController,
         updateFooterView()
     }
     
-    func updateFooterView() {
-        if let view = footerView {
-            let shouldShowFooter = topics.count > 0 && newsDataTask?.state == .running
-            view.isHidden = !shouldShowFooter
-        }
-    }
-    
-    func updateCollectionViewCellSize() {
-        guard let flowLayout = collectionView?.collectionViewLayout as? UICollectionViewFlowLayout,
-            let collectionViewSize = self.collectionView?.bounds.size else {
-                return
-        }
-        
-        if UIDevice.current.userInterfaceIdiom == .pad {
-            let columns: CGFloat = floor(collectionViewSize.width / 280.0)
-            let itemWidth = floor(((collectionViewSize.width - 20 - (columns-1)*10) / columns))
-            flowLayout.itemSize = CGSize(width: itemWidth, height: 235)
-        } else {
-            flowLayout.itemSize = CGSize(width: collectionViewSize.width - 20, height: 235)
-        }
-    }
-    
     override func viewDidLayoutSubviews() {
         super.viewDidLayoutSubviews()
         
@@ -170,50 +144,13 @@ class TrendingCardsViewController: UIViewController,
         }
     }
     
-    func setupReviewButtons() {
-        reviewButton.layer.borderColor = UIColor.white.cgColor
-        reviewButton.layer.borderWidth = 1
-        reviewButton.layer.cornerRadius = 8
-        
-        reviewCancelButton.layer.borderColor = UIColor.white.cgColor
-        reviewCancelButton.layer.borderWidth = 1
-        reviewCancelButton.layer.cornerRadius = 8
-    }
-    
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        guard let flowLayout = collectionView?.collectionViewLayout as? UICollectionViewFlowLayout else {
-            return
-        }
-        
-        flowLayout.minimumLineSpacing = 10
-        
-        if UIDevice.current.userInterfaceIdiom == .pad {
-            flowLayout.sectionInset = UIEdgeInsets(top: 10, left: 10, bottom: 10, right: 10)
-        }
-        
-        updateCollectionViewCellSize()
         
         setupReviewButtons()
         reviewView.isHidden = !self.reviewViewModel.shouldShowBanner()
         
-        let refreshCtrl = UIRefreshControl()
-        collectionView?.refreshControl = refreshCtrl
-        refreshCtrl.tintColor = UIColor(red: 0.99, green: 0.29, blue: 0.39, alpha: 1.00)
-        refreshCtrl.addTarget(self, action: #selector(fetchTrendingTopics), for: .valueChanged)
-        
-        //  Had to set content offset because of UIRefreshControl bug
-        //  http://stackoverflow.com/a/31224299/994129
-        collectionView?.contentOffset = CGPoint(x: 0, y: -refreshCtrl.frame.size.height)
-        
-        if let appDelegate = UIApplication.shared.delegate as? AppDelegate {
-            self.newsDataTask = appDelegate.newsDataTask
-            
-            if let n = appDelegate.newsFetched {
-                self.topics.append(contentsOf: n)
-            }
-        }
+        setupCollectionView()
         
         if topics.count == 0 {
             startRefreshing()
@@ -276,159 +213,11 @@ class TrendingCardsViewController: UIViewController,
     @IBAction func unwindFromSearch(segue: UIStoryboardSegue) {
     }
     
-    @IBAction func reviewButtonPressed(_ sender: Any) {
-        view.layoutIfNeeded()
-        reviewHeight.constant = 0
-        let animation = { [unowned self] in
-            self.reviewView.alpha = 0
-            self.view.layoutIfNeeded()
-        }
-        
-        let completion: (Bool) -> Void = { [unowned self] (completed) in
-            self.reviewView.isHidden = true
-            self.userSettingsManager.reviewDate = Date()
-            Answers.logCustomEvent(withName: "review_accept", customAttributes: nil)
-            SKStoreReviewController.requestReview()
-        }
-        
-        UIView.animate(withDuration: 0.3, animations: animation, completion: completion)
-    }
-    
-    @IBAction func reviewCancelButtonPressed(_ sender: Any) {
-        view.layoutIfNeeded()
-        reviewHeight.constant = 0
-        let animation = { [unowned self] in
-            self.reviewView.alpha = 0
-            self.view.layoutIfNeeded()
-        }
-        
-        let completion: (Bool) -> Void = { [unowned self] (completed) in
-            self.reviewView.isHidden = true
-            self.userSettingsManager.canceledReviewDate = Date()
-            Answers.logCustomEvent(withName: "review_cancel", customAttributes: nil)
-        }
-        
-        UIView.animate(withDuration: 0.3, animations: animation, completion: completion)
-    }
+}
 
-    // MARK: - UICollectionViewDataSource
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return self.topics.count
-    }
+// MARK: - TabbedViewController
+extension TrendingCardsViewController: TabbedViewController {
     
-    func collectionView(_ collectionView: UICollectionView,
-                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "cell", for: indexPath)
-            as? KeywordCollectionViewCell else {
-                
-            return UICollectionViewCell()
-        }
-        
-        let topic = topics[indexPath.row]
-        
-        guard let news = topic.news else {
-            return cell
-        }
-        
-        //  Tries to show a news on the topic card that contains an image,
-        //  it will show a news without an image if all the news of that topic don't
-        //  contain an image.
-        
-        let newsWithImages = news.filter({$0.imageUrl != nil})
-        
-        guard let firstNews = newsWithImages.first != nil ? newsWithImages.first : news.first else {
-            return cell
-        }
-        
-        if let newsDate = firstNews.date {
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateStyle = .short
-            dateFormatter.timeStyle = .none
-            dateFormatter.doesRelativeDateFormatting = true
-            cell.dateLabel.text = dateFormatter.string(from: newsDate)
-            
-            let timeFormatter = DateFormatter()
-            timeFormatter.dateStyle = .none
-            timeFormatter.timeStyle = .short
-            cell.timeLabel.text = timeFormatter.string(from: newsDate)
-        }
-
-        cell.titleLabel.text = topic.name
-        cell.bodyLabel.text = firstNews.title
-        cell.sourceLabel.text = firstNews.source
-        cell.newsQuantityLabel.text = "\(news.count) noticias"
-        cell.reactionLabel.text = topic.representativeReaction?.reaction ?? ""
-        
-        cell.imageView.contentMode = .center
-        
-        if let imgURL = firstNews.imageUrl {
-            cell.imageView.sd_setImage(
-                with: imgURL,
-                placeholderImage: UIImage(named: "icon_placeholder_big"),
-                options: [],
-                completed: { (_, error, _, _) in
-                    
-                    if error != nil {
-                        return
-                    }
-                    
-                    cell.imageView.contentMode = .scaleAspectFill
-            })
-        } else {
-            cell.imageView.image = nil
-        }
-        
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        let topic = topics[indexPath.row]
-        performSegue(withIdentifier: "news", sender: topic)
-        
-        Answers.logCustomEvent(
-            withName: "trending_item_tapped",
-            customAttributes: [
-                "topic": topic.name ?? "no_name"
-            ]
-        )
-    }
-    
-    func collectionView(_ collectionView: UICollectionView,
-                        viewForSupplementaryElementOfKind kind: String,
-                        at indexPath: IndexPath) -> UICollectionReusableView {
-        
-        guard let view = footerView else {
-            footerView = collectionView.dequeueReusableSupplementaryView(ofKind: kind,
-                                                                         withReuseIdentifier: "footer",
-                                                                         for: indexPath)
-            updateFooterView()
-            return footerView!
-        }
-        
-        return view
-    }
-    
-    // MARK: - UICollectionViewDelegate
-    func collectionView(_ collectionView: UICollectionView,
-                        willDisplay cell: UICollectionViewCell,
-                        forItemAt indexPath: IndexPath) {
-        
-        if indexPath.row == topics.count-1 {
-            let topic = topics[indexPath.row]
-            
-            guard let lastDate = topic.date else {
-                return
-            }
-            
-            let calendar = Calendar.current
-            let newDate = calendar.date(byAdding: .day, value: -1, to: lastDate)
-            
-            requestTrendingTopicsWithDate(newDate!)
-        }
-    }
-    
-    // MARK: - TabbedViewController
     func tabbedViewControllerWasDoubleTapped() {
         collectionView.setContentOffset(CGPoint.zero, animated: true)
     }
